requirements.txt
aniso8601==1.2.0
appdirs==1.4.2
babelfish==0.5.5
cffi==1.9.1
click==6.7
Flask==0.12
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.2
guessit==2.1.1
idna==2.4
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
marshmallow==2.13.1
packaging==16.8
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-slugify==1.2.1
pytz==2016.10
rebulk==0.8.2
six==1.10.0
SQLAlchemy==1.1.6
Unidecode==0.4.20
Werkzeug==0.11.15
